Add check for whether instance is actually locked.
An instance being present in the dataset SavedInstanceInfo did not
equal it actually being saved for that week. Therefore, it was
necessary to also parse the fifth value 'locked' returned from
GetSavedInstanceInfo(index).

In addition, this commit also includes a check for whether the
instance in question was actually found or if the for-loop
was fruitless.